changed to tsx for all components
